Aula 16.2 - Parte de criação (adição) de produtos

+++ api_scer/estatisticas/models.py
@@ -28,7 +28,7 @@ CLASSIFICACAO = [(item, item) for item in[
 class Produto(models.Model):
     nome = models.CharField(max_length=200, null=True)
     classificacao = models.CharField(max_length=200, choices=CLASSIFICACAO , null=True)
-    quantidade = models.PositiveIntegerField(null=True)
+    quantidade = models.DecimalField(max_digits=19, decimal_places=4, null=True)
     
     def __str__(self):
         return f'{self.nome}-{self.quantidade}'
